requirements.txt
python-telegram-bot==20.4
requests==2.31.0
python-dotenv==1.0.0
APScheduler==3.10.1
logging==0.4.9.6